Refactor getDependencyTreeOfRevision from AProxConnector into SCMConnector
