+++ communication/src/main/java/org/jboss/da/communication/aprox/api/AproxConnector.java
@@ -1,29 +1,14 @@
 package org.jboss.da.communication.aprox.api;
 
-import org.apache.maven.scm.ScmException;
 import org.jboss.da.communication.CommunicationException;
 import org.jboss.da.communication.aprox.model.GAVDependencyTree;
 import org.jboss.da.communication.model.GA;
 import org.jboss.da.communication.model.GAV;
-import org.jboss.da.communication.pom.PomAnalysisException;
 
 import java.util.List;
 import java.util.Optional;
 
 public interface AproxConnector {
-
-    /**
-     * Finds dependency trees of specific revision on scm url
-     * 
-     * @param scmUrl
-     * @param revision
-     * @param pomPath
-     * @return Optional of dependency tree of revision
-     * @throws PomAnalysisException When there is problem with the pom analysis
-     * @throws ScmException When checking out the repository failed
-     */
-    Optional<GAVDependencyTree> getDependencyTreeOfRevision(String scmUrl, String revision,
-            String pomPath) throws ScmException, PomAnalysisException;
 
     /**
      * Finds dependency trees of specific GAV
