+++ testsuite/src/test/java/org/jboss/da/test/server/communication/AproxRemoteTest.java
@@ -71,53 +71,4 @@ public class AproxRemoteTest {
         assertFalse(tree.isPresent());
     }
 
-    @Test
-    public void testGetDependencyTreeOfRevision() throws ScmException, PomAnalysisException {
-        String scmUrl = "https://github.com/project-ncl/dependency-analysis.git";
-
-        // commit id for version 0.3
-        String revision = "f34f4e1e";
-
-        // normal case
-        Optional<GAVDependencyTree> tree = aproxConnector.getDependencyTreeOfRevision(scmUrl,
-                revision, "");
-        assertTrue(tree.isPresent());
-
-        assertTrue(tree.get().getGav().equals(new GAV("org.jboss.da", "parent", "0.3.0")));
-        assertTrue(tree.get().getDependencies().isEmpty());
-
-        // with a slash in the pomPath
-        Optional<GAVDependencyTree> treeWithSlash = aproxConnector.getDependencyTreeOfRevision(
-                scmUrl, revision, "/");
-        assertTrue(treeWithSlash.isPresent());
-
-        assertTrue(treeWithSlash.get().getGav().equals(new GAV("org.jboss.da", "parent", "0.3.0")));
-        assertTrue(treeWithSlash.get().getDependencies().isEmpty());
-
-        // with application in the pomPath
-        Optional<GAVDependencyTree> treeApplication = aproxConnector.getDependencyTreeOfRevision(
-                scmUrl, revision, "application");
-        assertTrue(treeApplication.isPresent());
-
-        assertTrue(treeApplication.get().getGav()
-                .equals(new GAV("org.jboss.da", "application", "0.3.0")));
-        assertFalse(treeApplication.get().getDependencies().isEmpty());
-    }
-
-    @Test(expected = ScmException.class)
-    public void testDependencyTreeOfRevisionWrongRevision() throws ScmException,
-            PomAnalysisException {
-        String scmUrl = "https://github.com/project-ncl/does_not_exist.git";
-        Optional<GAVDependencyTree> tree = aproxConnector.getDependencyTreeOfRevision(scmUrl, "",
-                "");
-    }
-
-    @Test(expected = ScmException.class)
-    public void testDependencyTreeOfRevisionWrongRevision2() throws ScmException,
-            PomAnalysisException {
-        String scmUrl = "https://github.com/project-ncl/dependency-analysis.git";
-        String revision = "doesnotexist";
-        Optional<GAVDependencyTree> tree = aproxConnector.getDependencyTreeOfRevision(scmUrl,
-                revision, "");
-    }
 }
